requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.7
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.1
django-silk==5.0.3
djangorestframework==3.14.0
drf-spectacular==0.26.2
environ==1.0
gprof2dot==2022.7.29
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.28.2
rpds-py==0.9.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6